+++ configure.py
@@ -1407,6 +1407,7 @@ scylla_tools = ['tools/scylla-local-file-key-generator.cc',
                 'tools/scylla-types.cc',
                 'tools/scylla-sstable.cc',
                 'tools/scylla-nodetool.cc',
+                'tools/json_mutation_stream_parser.cc',
                 'tools/schema_loader.cc',
                 'tools/load_system_tablets.cc',
                 'tools/utils.cc',
